Refactor metricsx package
