+++ metricsx/middleware.go
@@ -24,42 +24,15 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
-	"net"
 	"net/http"
 	"net/url"
-	"runtime"
 	"strings"
-	"sync"
 	"time"
 
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-
-	"github.com/ory/x/cmdx"
-	"github.com/ory/x/resilience"
-
-	"github.com/pborman/uuid"
 	"github.com/segmentio/analytics-go"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/negroni"
 )
-
-var instance *Service
-var lock sync.Mutex
-
-// Service helps with providing context on metrics.
-type Service struct {
-	optOut bool
-	salt   string
-
-	o       *Options
-	context *analytics.Context
-
-	c analytics.Client
-	l logrus.FieldLogger
-
-	mem *MemoryStatistics
-}
 
 // Hash returns a hashed string of the value.
 func Hash(value string) string {
@@ -69,185 +42,6 @@ func Hash(value string) string {
 		panic(fmt.Sprintf("unable to hash value"))
 	}
 	return hex.EncodeToString(hash.Sum(nil))
-}
-
-// Options configures the metrics service.
-type Options struct {
-	// Service represents the service name, for example "ory-hydra".
-	Service string
-
-	// ClusterID represents the cluster id, typically a hash of some unique configuration properties.
-	ClusterID string
-
-	// IsDevelopment should be true if we assume that we're in a development environment.
-	IsDevelopment bool
-
-	// WriteKey is the segment API key.
-	WriteKey string
-
-	// WhitelistedPaths represents a list of paths that can be transmitted in clear text to segment.
-	WhitelistedPaths []string
-
-	// BuildVersion represents the build version.
-	BuildVersion string
-
-	// BuildHash represents the build git hash.
-	BuildHash string
-
-	// BuildTime represents the build time.
-	BuildTime string
-
-	// Config overrides the analytics.Config. If nil, sensible defaults will be used.
-	Config *analytics.Config
-
-	// MemoryInterval sets how often memory statistics should be transmitted. Defaults to every 12 hours.
-	MemoryInterval time.Duration
-}
-
-type void struct {
-}
-
-func (v *void) Logf(format string, args ...interface{}) {
-}
-
-func (v *void) Errorf(format string, args ...interface{}) {
-}
-
-// New returns a new metrics service. If one has been instantiated already, no new instance will be created.
-func New(
-	cmd *cobra.Command,
-	l logrus.FieldLogger,
-	o *Options,
-) *Service {
-	lock.Lock()
-	defer lock.Unlock()
-
-	if instance != nil {
-		return instance
-	}
-
-	if o.BuildTime == "" {
-		o.BuildTime = "unknown"
-	}
-
-	if o.BuildVersion == "" {
-		o.BuildVersion = "unknown"
-	}
-
-	if o.BuildHash == "" {
-		o.BuildHash = "unknown"
-	}
-
-	if o.Config == nil {
-		o.Config = &analytics.Config{
-			Interval:  time.Hour * 24,
-			BatchSize: 100,
-		}
-	}
-
-	o.Config.Logger = new(void)
-
-	if o.MemoryInterval < time.Minute {
-		o.MemoryInterval = time.Hour * 12
-	}
-
-	segment, err := analytics.NewWithConfig(o.WriteKey, *o.Config)
-	if err != nil {
-		l.WithError(err).Fatalf("Unable to initialise software quality assurance features.")
-		return nil
-	}
-
-	var oi analytics.OSInfo
-
-	optOut, err := cmd.Flags().GetBool("sqa-opt-out")
-	if !optOut {
-		optOut, err = cmd.Flags().GetBool("disable-telemetry")
-		if optOut {
-			l.Warn(`Command line argument "--disable-telemetry" has been deprecated and will be removed in an upcoming release. Use "--sqa-opt-out" instead.`)
-		}
-	}
-	cmdx.Must(err, "Unable to get command line flag.")
-
-	if !optOut {
-		optOut = viper.GetBool("sqa.opt_out")
-	}
-
-	if !optOut {
-		optOut = viper.GetBool("DISABLE_TELEMETRY")
-		if optOut {
-			l.Warn(`Environment variable "DISABLE_TELEMETRY" has been deprecated and will be removed in an upcoming release. Use configuration key "sqa.opt_out: true" or environment variable "SQA_OPT_OUT=true" instead.`)
-		}
-	}
-
-	if !optOut {
-		l.Info("Software quality assurance features are enabled. Learn more at: https://www.ory.sh/docs/ecosystem/sqa")
-		oi = analytics.OSInfo{
-			Version: fmt.Sprintf("%s-%s", runtime.GOOS, runtime.GOARCH),
-		}
-	}
-
-	m := &Service{
-		optOut: optOut,
-		salt:   uuid.New(),
-		o:      o,
-		c:      segment,
-		l:      l,
-		mem:    new(MemoryStatistics),
-		context: &analytics.Context{
-			IP: net.IPv4(0, 0, 0, 0),
-			App: analytics.AppInfo{
-				Name:    o.Service,
-				Version: o.BuildVersion,
-				Build:   fmt.Sprintf("%s/%s/%s", o.BuildVersion, o.BuildHash, o.BuildTime),
-			},
-			OS: oi,
-			Traits: analytics.NewTraits().
-				Set("optedOut", optOut).
-				Set("instanceId", uuid.New()).
-				Set("isDevelopment", o.IsDevelopment),
-			UserAgent: "github.com/ory/x/metricsx.Service/v0.0.1",
-		},
-	}
-
-	instance = m
-
-	go m.Identify()
-	go m.ObserveMemory()
-
-	return m
-}
-
-// Identify enables reporting to segment.
-func (sw *Service) Identify() {
-	if err := resilience.Retry(sw.l, time.Minute*5, time.Hour*24*30, func() error {
-		return sw.c.Enqueue(analytics.Identify{
-			UserId:  sw.o.ClusterID,
-			Traits:  sw.context.Traits,
-			Context: sw.context,
-		})
-	}); err != nil {
-		sw.l.WithError(err).Debug("Could not commit anonymized environment information")
-	}
-}
-
-// ObserveMemory commits memory statistics to segment.
-func (sw *Service) ObserveMemory() {
-	if sw.optOut {
-		return
-	}
-
-	for {
-		sw.mem.Update()
-		if err := sw.c.Enqueue(analytics.Track{
-			UserId:     sw.o.ClusterID,
-			Event:      "memstats",
-			Properties: analytics.Properties(sw.mem.ToMap()),
-			Context:    sw.context,
-		}); err != nil {
-			sw.l.WithError(err).Debug("Could not commit anonymized telemetry data")
-		}
-		time.Sleep(sw.o.MemoryInterval)
-	}
 }
 
 // ServeHTTP is a middleware for sending meta information to segment.
@@ -295,7 +89,17 @@ func anonymizeQuery(query url.Values, salt string) string {
 	return query.Encode()
 }
 
-func enqueue(rw http.ResponseWriter, r *http.Request, client analytics.Client, log logrus.FieldLogger, whitelist []string, latency time.Duration, salt string, userID string, ctx *analytics.Context) {
+func enqueue(
+	rw http.ResponseWriter,
+	r *http.Request,
+	client analytics.Client,
+	log logrus.FieldLogger,
+	whitelist []string,
+	latency time.Duration,
+	salt string,
+	userID string,
+	ctx *analytics.Context,
+) {
 	scheme := "https:"
 	if r.TLS == nil {
 		scheme = "http:"
