+++ metricsx/prometheus.go
@@ -1,27 +1,226 @@
 package metricsx
 
 import (
+	"fmt"
+	"net"
+	"runtime"
+	"sync"
+	"time"
+
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+
+	"github.com/ory/x/cmdx"
+	"github.com/ory/x/resilience"
 	"github.com/pborman/uuid"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/segmentio/analytics-go"
-
-	"github.com/ory/hydra/client"
-	"github.com/ory/hydra/consent"
-	"github.com/ory/hydra/driver"
-	"github.com/ory/hydra/jwk"
-	"github.com/ory/hydra/oauth2"
+	"github.com/sirupsen/logrus"
 )
 
-var (
-	segmentOptions *metricsx.SegmentOptions
-)
+var lock sync.Mutex
+var segmentInstance *Service
+var DefaultWatchDuration = time.Hour * 12
+
+type void struct {
+}
+
+func (v *void) Logf(format string, args ...interface{}) {
+}
+
+func (v *void) Errorf(format string, args ...interface{}) {
+}
+
+// Options configures the metrics service.
+type Options struct {
+	// Service represents the service name, for example "ory-hydra".
+	Service string
+
+	// ClusterID represents the cluster id, typically a hash of some unique configuration properties.
+	ClusterID string
+
+	// IsDevelopment should be true if we assume that we're in a development environment.
+	IsDevelopment bool
+
+	// WriteKey is the segment API key.
+	WriteKey string
+
+	// WhitelistedPaths represents a list of paths that can be transmitted in clear text to segment.
+	WhitelistedPaths []string
+
+	// BuildVersion represents the build version.
+	BuildVersion string
+
+	// BuildHash represents the build git hash.
+	BuildHash string
+
+	// BuildTime represents the build time.
+	BuildTime string
+
+	// Config overrides the analytics.Config. If nil, sensible defaults will be used.
+	Config *analytics.Config
+
+	// MemoryInterval sets how often memory statistics should be transmitted. Defaults to every 12 hours.
+	MemoryInterval time.Duration
+}
+
+// The Observer type is used for populating data into prometheus metrics that are repeatedly pulled from the database.
+type Observer interface {
+	Observe() error
+}
+
+// Watch will continuously call o.Observe every time `i` ticks.
+func Watch(o Observer, i time.Duration, log logrus.FieldLogger) {
+	t := time.NewTicker(i)
+
+	for {
+		select {
+		case <-t.C:
+			if err := o.Observe(); err != nil {
+				log.WithError(err).Infoln("Error when gathering metrics")
+			}
+		}
+	}
+}
+
+// Service helps with providing context on metrics.
+type Service struct {
+	optOut bool
+	salt   string
+
+	o       *Options
+	context *analytics.Context
+
+	c analytics.Client
+	l logrus.FieldLogger
+
+	mem *MemoryStatistics
+}
+
+// ObserveMemory commits memory statistics to segment.
+func (sw *Service) ObserveMemory() {
+	if sw.optOut {
+		return
+	}
+
+	for {
+		sw.mem.Update()
+		if err := sw.c.Enqueue(analytics.Track{
+			UserId:     sw.o.ClusterID,
+			Event:      "memstats",
+			Properties: analytics.Properties(sw.mem.ToMap()),
+			Context:    sw.context,
+		}); err != nil {
+			sw.l.WithError(err).Debug("Could not commit anonymized telemetry data")
+		}
+		time.Sleep(sw.o.MemoryInterval)
+	}
+}
+
+// New returns a new metrics service. If one has been instantiated already, no new instance will be created.
+func New(
+	cmd *cobra.Command,
+	l logrus.FieldLogger,
+	o *Options,
+) *Service {
+	lock.Lock()
+	defer lock.Unlock()
+
+	if segmentInstance != nil {
+		return segmentInstance
+	}
+
+	if o.BuildTime == "" {
+		o.BuildTime = "unknown"
+	}
+
+	if o.BuildVersion == "" {
+		o.BuildVersion = "unknown"
+	}
+
+	if o.BuildHash == "" {
+		o.BuildHash = "unknown"
+	}
+
+	if o.Config == nil {
+		o.Config = &analytics.Config{
+			Interval:  time.Hour * 24,
+			BatchSize: 100,
+		}
+	}
+
+	o.Config.Logger = new(void)
+
+	if o.MemoryInterval < time.Minute {
+		o.MemoryInterval = DefaultWatchDuration
+	}
+
+	segment, err := analytics.NewWithConfig(o.WriteKey, *o.Config)
+	if err != nil {
+		l.WithError(err).Fatalf("Unable to initialise software quality assurance features.")
+		return nil
+	}
+
+	var oi analytics.OSInfo
+
+	optOut := isOptedOut(cmd, l)
+
+	if !optOut {
+		l.Info("Software quality assurance features are enabled. Learn more at: https://www.ory.sh/docs/ecosystem/sqa")
+		oi = analytics.OSInfo{
+			Version: fmt.Sprintf("%s-%s", runtime.GOOS, runtime.GOARCH),
+		}
+	}
+
+	m := &Service{
+		optOut: optOut,
+		salt:   uuid.New(),
+		o:      o,
+		c:      segment,
+		l:      l,
+		mem:    new(MemoryStatistics),
+		context: &analytics.Context{
+			IP: net.IPv4(0, 0, 0, 0),
+			App: analytics.AppInfo{
+				Name:    o.Service,
+				Version: o.BuildVersion,
+				Build:   fmt.Sprintf("%s/%s/%s", o.BuildVersion, o.BuildHash, o.BuildTime),
+			},
+			OS: oi,
+			Traits: analytics.NewTraits().
+				Set("optedOut", optOut).
+				Set("instanceId", uuid.New()).
+				Set("isDevelopment", o.IsDevelopment),
+			UserAgent: "github.com/ory/x/metricsx.Service/v0.0.1",
+		},
+	}
+
+	segmentInstance = m
+
+	go m.Identify()
+	go m.ObserveMemory()
+
+	return m
+}
+
+// Identify identifies a unique user to Segment
+func (sw *Service) Identify() {
+	if err := resilience.Retry(sw.l, time.Minute*5, time.Hour*24*30, func() error {
+		return sw.c.Enqueue(analytics.Identify{
+			UserId:  sw.o.ClusterID,
+			Traits:  sw.context.Traits,
+			Context: sw.context,
+		})
+	}); err != nil {
+		sw.l.WithError(err).Debug("Could not commit anonymized environment information")
+	}
+}
 
 // GaugeVec is an interface which prometheus.GaugeVec implements.
 // This allows for more generic wrapping of prometheus.GaugeVec
 type GaugeVec interface {
 	prometheus.Collector
 	With(prometheus.Labels) prometheus.Gauge
-	WithLabelValues(...string) prometheus.Gauge
 }
 
 // CounterVec is an interface which prometheus.CounterVec implements.
@@ -29,40 +228,41 @@ type GaugeVec interface {
 type CounterVec interface {
 	prometheus.Collector
 	With(prometheus.Labels) prometheus.Counter
-	WithLabelValues(...string) prometheus.Counter
 }
 
 // NewGaugeVec returns a GaugeVec. This function uses options provided to the process to create
 // the appropriate wrappers.
+// As of writing this, only Segment is available. However, in the future, it is likely that many different services could wrap the metric.
 func NewGaugeVec(opts prometheus.GaugeOpts, labels []string) GaugeVec {
-	var c GaugeVec = prometheus.NewGaugeVec(opts, labels)
+	return WithSegmentGaugeVec(opts, prometheus.NewGaugeVec(opts, labels))
 
-	if segmentOptions != nil {
-		c = WithSegmentGaugeVec(c)
-	}
+}
 
-	return c
+// NewGauge creates a GaugeVec, but with an empty list of labels (using With).
+// When using NewGauge, it is perfectly safe (and preferred) to use a prometheus.Gauge type.
+// As of writing this, only Segment is available. However, in the future, it is likely that many different services could wrap the metric.
+func NewGauge(opts prometheus.GaugeOpts) prometheus.Gauge {
+	return NewGaugeVec(opts, []string{}).With(prometheus.Labels{})
 }
 
 // NewCounterVec returns a CounterVec. This function uses options provided to the process to create
 // the appropriate wrappers.
+// As of writing this, only Segment is available. However, in the future, it is likely that many different services could wrap the metric.
 func NewCounterVec(opts prometheus.CounterOpts, labels []string) CounterVec {
-	var c CounterVec = prometheus.NewCounterVec(opts, labels)
-
-	if segmentOptions != nil {
-		c = WithSegmentCounterVec(c)
-	}
-
-	return c
+	return WithSegmentCounterVec(opts, prometheus.NewCounterVec(opts, labels))
 }
 
-// Setup will initialize the available options
-func Setup(d driver.Driver) {
-	var (
-		l   = d.Registry().Logger()
-		ctx = context.Background()
-	)
+// NewCounter creates a CounterVec, but with an empty list of labels (using With).
+// When using NewCounter, it is perfectly safe (and preferred) to use a prometheus.Counter type.
+// As of writing this, only Segment is available. However, in the future, it is likely that many different services could wrap the metric.
+func NewCounter(opts prometheus.CounterOpts) prometheus.Counter {
+	return NewCounterVec(opts, []string{}).With(prometheus.Labels{})
+}
 
+func isOptedOut(
+	cmd *cobra.Command,
+	l logrus.FieldLogger,
+) bool {
 	optOut, err := cmd.Flags().GetBool("sqa-opt-out")
 	if !optOut {
 		optOut, err = cmd.Flags().GetBool("disable-telemetry")
@@ -83,60 +283,5 @@ func Setup(d driver.Driver) {
 		}
 	}
 
-	if !optOut {
-		l.Info("Software quality assurance features are enabled. Learn more at: https://www.ory.sh/docs/ecosystem/sqa")
-		segmentOptions = &metricsx.SegmentOptions{
-			Service: "ory-hydra",
-			ClusterID: metricsx.Hash(fmt.Sprintf("%s|%s",
-				d.Configuration().IssuerURL().String(),
-				d.Configuration().DSN(),
-			)),
-			IsDevelopment: d.Configuration().DSN() == "memory" ||
-				d.Configuration().IssuerURL().String() == "" ||
-				strings.Contains(d.Configuration().IssuerURL().String(), "localhost"),
-			WhitelistedPaths: []string{
-				jwk.KeyHandlerPath,
-				jwk.WellKnownKeysPath,
-
-				client.ClientsHandlerPath,
-
-				oauth2.DefaultConsentPath,
-				oauth2.DefaultLoginPath,
-				oauth2.DefaultPostLogoutPath,
-				oauth2.DefaultLogoutPath,
-				oauth2.DefaultErrorPath,
-				oauth2.TokenPath,
-				oauth2.AuthPath,
-				oauth2.LogoutPath,
-				oauth2.UserinfoPath,
-				oauth2.WellKnownPath,
-				oauth2.JWKPath,
-				oauth2.IntrospectPath,
-				oauth2.RevocationPath,
-				oauth2.FlushPath,
-
-				consent.ConsentPath,
-				consent.ConsentPath + "/accept",
-				consent.ConsentPath + "/reject",
-				consent.LoginPath,
-				consent.LoginPath + "/accept",
-				consent.LoginPath + "/reject",
-				consent.LogoutPath,
-				consent.LogoutPath + "/accept",
-				consent.LogoutPath + "/reject",
-				consent.SessionsPath + "/login",
-				consent.SessionsPath + "/consent",
-
-				healthx.AliveCheckPath,
-				healthx.ReadyCheckPath,
-				healthx.VersionPath,
-				driver.MetricsPrometheusPath,
-				"/",
-			},
-			WriteKey:     "h8dRH3kVCWKkIFWydBmWsyYHR4M0u0vr",
-			BuildVersion: d.Registry().BuildVersion(),
-			BuildTime:    d.Registry().BuildDate(),
-			BuildHash:    d.Registry().BuildHash(),
-		}
-	}
+	return optOut
 }
